garmin_data_pull/requirements.txt
numpy
matplotlib
pandas
sqlalchemy
psycopg2
garminconnect
python-dotenv